Managed to compile a ROM!
There were some BY HAND changes. So, the tool is far from understood. I think that each of these pieces (about a dozen tools for this toolchain) will take time to understand. But, compiling a simple ASM file seems ok. The pieces will need an IDE - which is what I'm after with this program. While doing this setup - there would be a viable text-editor based IDE with the base compilers setup for the user. That will take some simple fussiness with paths / toolchain commands / file validation ... blah blah blah. It may not bear fruit until we really dig in and just read the Pretendo docs.